+++ src/Exceptions/AttributeCasterException.php
@@ -5,15 +5,8 @@ namespace Somnambulist\Components\AttributeModel\Exceptions;
 use Exception;
 use function sprintf;
 
-/**
- * Class AttributeCasterException
- *
- * @package    Somnambulist\Components\AttributeModel\Exceptions
- * @subpackage Somnambulist\Components\AttributeModel\Exceptions\AttributeCasterException
- */
 class AttributeCasterException extends Exception
 {
-
     public static function missingTypeFor(string $type): self
     {
         return new self(sprintf('Missing type caster for "%s"', $type));