+++ src/TypeCasters/ExternalIdentityCaster.php
@@ -5,26 +5,14 @@ namespace Somnambulist\Components\AttributeModel\TypeCasters;
 use Somnambulist\Components\AttributeModel\Contracts\AttributeCasterInterface;
 use Somnambulist\Components\Domain\Entities\Types\Identity\ExternalIdentity;
 
-/**
- * Class ExternalIdentityCaster
- *
- * @package    Somnambulist\Components\AttributeModel\TypeCasters
- * @subpackage Somnambulist\Components\AttributeModel\TypeCasters\ExternalIdentityCaster
- */
 final class ExternalIdentityCaster implements AttributeCasterInterface
 {
-
-    private string $providerAttribute;
-    private string $identityAttribute;
-    private bool $remove;
-    private ?array $types;
-
-    public function __construct(string $providerAttribute = 'provider', string $identityAttribute = 'identity', bool $remove = true, ?array $types = null)
-    {
-        $this->providerAttribute = $providerAttribute;
-        $this->identityAttribute = $identityAttribute;
-        $this->remove            = $remove;
-        $this->types             = $types;
+    public function __construct(
+        private string $providerAttribute = 'provider',
+        private string $identityAttribute = 'identity',
+        private bool $remove = true,
+        private ?array $types = null
+    ) {
     }
 
     public function types(): array
@@ -37,7 +25,7 @@ final class ExternalIdentityCaster implements AttributeCasterInterface
         return in_array($type, $this->types());
     }
 
-    public function cast(array &$attributes, $attribute, string $type): void
+    public function cast(array &$attributes, mixed $attribute, string $type): void
     {
         if (!isset($attributes[$this->providerAttribute], $attributes[$this->identityAttribute])) {
             return;