+++ tests/Support/Stubs/Models/MyEnum.php
@@ -4,15 +4,8 @@ namespace Somnambulist\Components\AttributeModel\Tests\Support\Stubs\Models;
 
 use Somnambulist\Components\Domain\Entities\AbstractEnumeration;
 
-/**
- * Class MyEnum
- *
- * @package    Somnambulist\Components\AttributeModel\Tests\Support\Stubs\Models
- * @subpackage Somnambulist\Components\AttributeModel\Tests\Support\Stubs\Models\MyEnum
- */
 class MyEnum extends AbstractEnumeration
 {
-
     const VAL_1 = 1;
     const VAL_2 = 2;
     const VAL_3 = 3;